FROMLIST: arm64: kvm: stop treating register x18 as caller save
In preparation of using x18 as a task struct pointer register when
running in the kernel, stop treating it as caller save in the KVM
guest entry/exit code. Currently, the code assumes there is no need
to preserve it for the host, given that it would have been assumed
clobbered anyway by the function call to __guest_enter(). Instead,
preserve its value and restore it upon return.

Bug: 112277034
Change-Id: Ibaf62ff542a458180a2538206f1b11bf2c96c19b
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Signed-off-by: karthik558 <karthik.lal558@gmail.com>